Add 25/3/2022

+++ recursive algorithm - fibonacci.cpp
@@ -1,11 +1,11 @@
 #include <bits/stdc++.h>
 using namespace std;
-long long Fibo(int n){
+long long Fibo(long long n){
     if(n==1||n==2) return 1;
     return Fibo(n-1)+Fibo(n-2);
 }
 int main(){
-	int a;
+	long long a;
 	cin >> a;
 	cout << Fibo(a);
 	return 0;
